Handle queries with joined tables
When l10n modifies a query, it adds conditions on the primary key,
language_code and deleted_at, without specifying the table name.

If the query does a joins (to do an order by on a related table for
example), it can lead to ambiguous queries, because the column names
may exist in the joined tables.

This patch adds the needed table names in the modified query, so that
there is no more ambiguous column name.